requirements.txt
requests==2.31.0
python-dateutil==2.8.2
pandas==2.1.1
pytz==2023.3 